Move get_scores_row_inference to leaf module to break export import cycle.
exports.py imports inference entry at module level instead of using a lazy import through scores/__init__.py.
